Move array_poolt class to its own files
This makes it easier to find array_poolt related code.
