+++ src/solvers/Makefile
@@ -159,6 +159,7 @@ SRC = $(BOOLEFORCE_SRC) \
       refinement/bv_refinement_loop.cpp \
       refinement/refine_arithmetic.cpp \
       refinement/refine_arrays.cpp \
+      strings/array_pool.cpp \
       strings/string_builtin_function.cpp \
       strings/string_refinement.cpp \
       strings/string_refinement_util.cpp \
