+++ src/solvers/strings/string_constraint_generator_main.cpp
@@ -35,23 +35,6 @@ string_constraint_generatort::string_constraint_generatort(const namespacet &ns)
 {
 }
 
-/// generate a new symbol expression of the given type with some prefix
-/// \par parameters: a prefix and a type
-/// \return a symbol of type tp whose name starts with "string_refinement#"
-///   followed by prefix
-symbol_exprt symbol_generatort::
-operator()(const irep_idt &prefix, const typet &type)
-{
-  std::ostringstream buf;
-  buf << "string_refinement#" << prefix << "#" << ++symbol_count;
-  irep_idt name(buf.str());
-  symbol_exprt result(name, type);
-#ifdef DEBUG
-  created_symbols.push_back(result);
-#endif
-  return result;
-}
-
 exprt sum_overflows(const plus_exprt &sum)
 {
   PRECONDITION(sum.operands().size() == 2);
@@ -65,104 +48,6 @@ exprt sum_overflows(const plus_exprt &sum)
   return and_exprt(
     equal_exprt(op0_negative, op1_negative),
     notequal_exprt(op1_negative, sum_negative));
-}
-
-/// Associate an actual finite length to infinite arrays
-/// \param s: array expression representing a string
-/// \return expression for the length of `s`
-exprt array_poolt::get_length(const array_string_exprt &s) const
-{
-  if(s.length() == infinity_exprt(s.length().type()))
-  {
-    auto it = length_of_array.find(s);
-    if(it != length_of_array.end())
-      return it->second;
-  }
-  return s.length();
-}
-
-/// construct a string expression whose length and content are new variables
-/// \par parameters: a type for string
-/// \return a string expression
-array_string_exprt
-array_poolt::fresh_string(const typet &index_type, const typet &char_type)
-{
-  symbol_exprt length = fresh_symbol("string_length", index_type);
-  array_typet array_type(char_type, length);
-  symbol_exprt content = fresh_symbol("string_content", array_type);
-  array_string_exprt str = to_array_string_expr(content);
-  created_strings_value.insert(str);
-  return str;
-}
-
-// Make a new char array for a char pointer. The size of the char array is a
-// variable with no constraint.
-array_string_exprt array_poolt::make_char_array_for_char_pointer(
-  const exprt &char_pointer,
-  const typet &char_array_type)
-{
-  PRECONDITION(char_pointer.type().id() == ID_pointer);
-  PRECONDITION(char_array_type.id() == ID_array);
-  PRECONDITION(
-    char_array_type.subtype().id() == ID_unsignedbv ||
-    char_array_type.subtype().id() == ID_signedbv);
-
-  if(char_pointer.id() == ID_if)
-  {
-    const if_exprt &if_expr = to_if_expr(char_pointer);
-    const array_string_exprt t =
-      make_char_array_for_char_pointer(if_expr.true_case(), char_array_type);
-    const array_string_exprt f =
-      make_char_array_for_char_pointer(if_expr.false_case(), char_array_type);
-    const array_typet array_type(
-      char_array_type.subtype(),
-      if_exprt(
-        if_expr.cond(),
-        to_array_type(t.type()).size(),
-        to_array_type(f.type()).size()));
-    // BEWARE: this expression is possibly type-inconsistent and must not be
-    // used for any purposes other than passing it to concretisation
-    return to_array_string_expr(if_exprt(if_expr.cond(), t, f, array_type));
-  }
-  const bool is_constant_array =
-    char_pointer.id() == ID_address_of &&
-    (to_address_of_expr(char_pointer).object().id() == ID_index) &&
-    char_pointer.op0().op0().id() == ID_array;
-  if(is_constant_array)
-  {
-    return to_array_string_expr(
-      to_index_expr(to_address_of_expr(char_pointer).object()).array());
-  }
-  const std::string symbol_name = "char_array_" + id2string(char_pointer.id());
-  const auto array_sym =
-    to_array_string_expr(fresh_symbol(symbol_name, char_array_type));
-  const auto insert_result =
-    arrays_of_pointers.insert({char_pointer, array_sym});
-  return to_array_string_expr(insert_result.first->second);
-}
-
-void array_poolt::insert(
-  const exprt &pointer_expr,
-  array_string_exprt &array_expr)
-{
-  const exprt &length = array_expr.length();
-  if(length == infinity_exprt(length.type()))
-  {
-    auto pair = length_of_array.insert(
-      std::make_pair(array_expr, fresh_symbol("string_length", length.type())));
-    array_expr.length() = pair.first->second;
-  }
-
-  const auto it_bool =
-    arrays_of_pointers.insert(std::make_pair(pointer_expr, array_expr));
-  created_strings_value.insert(array_expr);
-  INVARIANT(
-    it_bool.second, "should not associate two arrays to the same pointer");
-}
-
-const std::set<array_string_exprt> &array_poolt::created_strings() const
-{
-  return created_strings_value;
 }
 
 /// Associate a char array to a char pointer.
@@ -203,18 +88,6 @@ exprt string_constraint_generatort::associate_length_to_array(
   const auto &length = array_pool.get_length(array_expr);
   constraints.existential.push_back(equal_exprt(length, new_length));
   return from_integer(0, f.type());
-}
-
-/// casts an expression to a string expression, or fetches the actual
-/// string_exprt in the case of a symbol.
-/// \param pool: pool of arrays representing strings
-/// \param expr: an expression of refined string type
-/// \return a string expression
-array_string_exprt get_string_expr(array_poolt &pool, const exprt &expr)
-{
-  PRECONDITION(is_refined_string_type(expr.type()));
-  const refined_string_exprt &str = to_string_expr(expr);
-  return char_array_of_pointer(pool, str.content(), str.length());
 }
 
 void string_constraintst::clear()
@@ -309,34 +182,6 @@ std::pair<exprt, string_constraintst> add_axioms_for_constrain_characters(
   auto constraints = add_constraint_on_characters(
     fresh_symbol, s, start, end, char_set_string.c_str());
   return {from_integer(0, get_return_code_type()), std::move(constraints)};
-}
-
-/// Creates a new array if the pointer is not pointing to an array
-/// \todo This should be replaced by make_char_array_for_char_pointer
-const array_string_exprt &
-array_poolt::find(const exprt &pointer, const exprt &length)
-{
-  const array_typet array_type(pointer.type().subtype(), length);
-  const array_string_exprt array =
-    make_char_array_for_char_pointer(pointer, array_type);
-  return *created_strings_value.insert(array).first;
-}
-
-/// Adds creates a new array if it does not already exists
-/// \todo This should be replaced
-/// by array_poolt.make_char_array_for_char_pointer
-const array_string_exprt &char_array_of_pointer(
-  array_poolt &pool,
-  const exprt &pointer,
-  const exprt &length)
-{
-  return pool.find(pointer, length);
-}
-
-array_string_exprt of_argument(array_poolt &array_pool, const exprt &arg)
-{
-  const auto string_argument = expr_checked_cast<struct_exprt>(arg);
-  return array_pool.find(string_argument.op1(), string_argument.op0());
 }
 
 signedbv_typet get_return_code_type()
